Improve const-correctness and config usage across codebase
- Refactored several member functions to be const noexcept.
- Renamed AlienFormation to AlienFormationConfig and made its members static constexpr, updating all related usage.
- Moved alien shooting cooldown logic into the Game class.
- Updated Wall and Window to use gsl::narrow_cast and static constexpr members, respectively.
- Marked relevant lambdas as noexcept and fixed minor typos in comments.

+++ Source/Wall.h
@@ -34,6 +34,7 @@ public:
 			static_cast<int>(position_.y - 100.0f),
 			WHITE);
 
-		DrawText(TextFormat("%i", health_), position_.x - 21.0f, position_.y + 10.0f, 40, RED);
+		DrawText(TextFormat("%i", health_), gsl::narrow_cast<int>(position_.x - 21.0f), 
+											gsl::narrow_cast<int>(position_.y + 10.0f), 40, RED);
 	}
 };
